main/requirements.txt
pycryptodome==3.11.0
requests==2.26.0
geopy
